Enable Sound editor for all platforms...
Unknown audio encoding will simply disable the play/import/export buttons, but it is still possible to see the list of sounds (and drag-drop)

For the Wii platform, sounds from OG+Spyro are now fully supported. (It turned out it was uncompressed PCM but in big endian, like on 360/PS3.)
Alice + HTTYD use Nintendo's ADPCM, so unsupported for now.